Rename Rekor feeder, remove fetching stable checkpoint
Renaming this feeder to distinguish between this feeder and the upcoming
Rekor v2 feeder which will simply use the tiles feeder.

Also remove fetching the stable checkpoint, since this has never
actually been integrated into any workflow in practice.
